PropertyInput Uses Your HTML
Now, instead of using a bunch of carry-over attributes, you should try
to supply the html of the input you wish you to use to associate with
the related component property.  If you don't provide any HTML, a
generic input will be used (and can then be customized), but it's better
practice for your CARD, the master of its internal content, to be in
charge of how that should work.