PropertyObserver Uses InnerText As Bound Function Script
Now when you create a property observer, you specify the property to
observe with the "property" attribute and write your callback in the
body of the element.

The element's callback is memoized so it cannot be changed after it's
been created the first time.

The callback is bound to the context of the Card and has access to the
following scoped variables,
- value - the new property value
- component - The Custom Element which wraps the component class.

Much Nicer!