ScriptPart elements are context scoped script nodes.
When building out your app, your card may need to respond to component
related events and actions.  For example, your card may need to respond
to click events in an element on a component for recording analytics or
your card may need to trigger an effect when something happens to the
component itself such as property changes or child list updates.

OnEvent allows you to provide a list of events that you would like to
listen to on the elements parent node. The event and component are
scoped into the function, which is bound to the card's context ('this'
is the parent card.)