Better Handling of Text Mutations
Chrome and Firefox are slightly different in how they handle text data
node changes.

In Chrome, you can observe the text node and use innerText on the parent
node to trigger the characterData mutation.

In FF, the innerText method replaces the text and discards the mutation
effect.

In FF, you either have to observe the text node itself, then use the
data setters. OR, observe the parent node and wait for childList changes
in the subtree to determine if a textNode has changed.

So, now Observation checks the target node to try and properly handle
text mutations.  If the target node is a text node, than it will listen
for mutations on the text nodes data element.  If it's a dom node, it
will use the childList and subtree options then, from the mutation
arguments, determines whether or not it's actually a text mutation or a
child list mutation.

If it's a simple dom node, such as a span, than it should be ok to use
the CharacterData mutation, targeting either the node or the textNode.
However, if you plan on changing the inner html of a node in a way which
puts the text into somewhere less predicatable, you will want to use the
childlist-subtree mutation observation instead and handle checking for
text changes in your callback.