NavMap Component wraps a nav node list
Reactive navigation list - as you add links the nav node re-renders.
You can add a 'useevent' attribute to disable the normal functionality
of the anchor tag and dispatch a custom event instead.

That way, if you are using an ajax loader or need to perform some task
before redirecting the user, you can easily do so.  Either use an
on-event part or attach an event listener in your js.

The navmap attribute of the custom element looks like this: 'label:url,
label: url', ...'