Merge CardComponent with Component
The original idea was that there would be components that didn't belong
to a card. That concept was renamed to Parts earlier.  There's no longer
a need to have a CardComponent class separate from the Component class,
so it's been removed.

Also - since all components are Components for Cards, the 'cards'
subdirectory of Components has been renamed to 'generics'.  It
represents a collection of generic Components which can be used /
extended to add standardized functionality to your site.