Component content container custom elements
In order to ensure only content gets rewritten on attribute changes,
we've added 2 generic custom elements.  component-content and
component-parts.  Each has the relevant slot attribute to ensure it's
content is treated appropriately.

Updated Part to ensure the component it thinks is the parent component
is actually a component and not another custom-element.