Add ability for OnEvent to take targets beyond components.
In order to improve Card rendering, it's important that some parts can
attach to normal html such as buttons and image tags, without requiring
a bunch of wrappers. The 'target' attribute on on-event elements allow
you to specify 2 custom options.  Omitting the attribute uses a parent
component by default. The usable options are 'parent' and a css selector
which is queryied off of the parent element of the part.