Handle editable properties
You can now make a component edit it's own property via contenteditable.
While the component element is being edited, the matching attributes
and component property are updated.  However, the component won't
re-render until the editable has been blurred.

Use a property input part when editing isn't the default nature of the
component, but is being hooked in by the parent card.

Use a property observer to connect a component that reacts to the
property change, instead of tying that into the same component being
edited.